Reorder and Refactor
