+++ app/src/main/java/com/undef/localhandsbrambillafunes/ui/screens/productdetail/ProductDetailScreen.kt
@@ -101,15 +101,15 @@ fun ProductDetailScreen(navController: NavController, product: Product) {
                 ),
                 actions = {
                     // Botón para ir a Favoritos
-                    IconButton(onClick = { navController.navigate(route = AppScreens.FavoritesScreen.route) }) {
+                    IconButton(onClick = { navController.navigate(route = AppScreens.SearchBarScreen.route) }) {
                         Icon(
-                            Icons.Filled.Favorite,
-                            contentDescription = "Seccion de Favoritos"
+                            Icons.Filled.Search,
+                            contentDescription = "Buscar"
                         )
                     }
 
                     // Botón para ir a Perfil
-                    IconButton(onClick = { /* TODO: Implementar navegación */ }) {
+                    IconButton(onClick = { navController.navigate(AppScreens.ProfileScreen.route) }) {
                         Icon(
                             Icons.Filled.Person,
                             contentDescription = "Seccion de Perfil"
@@ -151,15 +151,15 @@ fun ProductDetailScreen(navController: NavController, product: Product) {
                     label = { Text("Inicio")},
                     colors = navBarItemColors,
                     selected = true,
-                    onClick = { navController.navigate(route = AppScreens.LocalHandsApp.route) }
+                    onClick = { navController.navigate(route = AppScreens.HomeScreen.route) }
                 )
-                // Boton de explorar o buscar
+                // Boton de Favoritos
                 NavigationBarItem(
-                    icon = { Icon(Icons.Filled.Search, contentDescription = "Buscar") },
-                    label = { Text("Buscar")},
+                    icon = { Icon(Icons.Filled.Favorite, contentDescription = "Favoritos") },
+                    label = { Text("Favoritos")},
                     colors = navBarItemColors,
                     selected = true,
-                    onClick = { /* TODO: Implementar Busqueda */ }
+                    onClick = { navController.navigate(AppScreens.FavoritesScreen.route)}
                 )
                 // Boton para vender
                 NavigationBarItem(
@@ -175,7 +175,7 @@ fun ProductDetailScreen(navController: NavController, product: Product) {
                     label = { Text("Categorias")},
                     colors = navBarItemColors,
                     selected = true,
-                    onClick = { /* TODO: Implementar navegacion */ }
+                    onClick = { navController.navigate(AppScreens.CategoryScreen.route) }
                 )
             }
         }
