+++ app/src/main/java/com/undef/localhandsbrambillafunes/ui/screens/profile/ProfileScreen.kt
@@ -1,6 +1,5 @@
 package com.undef.localhandsbrambillafunes.ui.screens.profile
 
-import android.app.AlertDialog
 import android.net.Uri
 import android.util.Patterns
 import android.widget.Toast
@@ -10,6 +9,7 @@ import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -21,16 +21,28 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.AccountCircle
+import androidx.compose.material.icons.filled.ArrowBackIosNew
+import androidx.compose.material.icons.filled.Favorite
+import androidx.compose.material.icons.filled.Home
+import androidx.compose.material.icons.filled.Menu
+import androidx.compose.material.icons.filled.Settings
+import androidx.compose.material.icons.filled.Shop
 import androidx.compose.material.icons.filled.Visibility
 import androidx.compose.material.icons.filled.VisibilityOff
 import androidx.compose.material3.Button
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.NavigationBar
+import androidx.compose.material3.NavigationBarItem
+import androidx.compose.material3.NavigationBarItemDefaults
 import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -41,15 +53,14 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import coil.compose.rememberAsyncImagePainter
-import com.undef.localhandsbrambillafunes.R
 import com.undef.localhandsbrambillafunes.ui.navigation.AppScreens
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun ProfileScreen(navController: NavController) {
     //Necesario para crear los Toast
@@ -91,125 +102,227 @@ fun ProfileScreen(navController: NavController) {
         imageUri = uri // Guarda la URI seleccionada
     }
 
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp)
-            .verticalScroll(rememberScrollState()),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        // Imagen de perfil clickeable
-        Box(modifier = Modifier
-            .size(100.dp)
-            .clip(CircleShape) //Recorta el contenido de la box formando un circulo
-            .clickable {
-                //Solo mostrar archivos que comiencen con image/, como image/jpeg, image/png, etc. (de tipo MIME)
-                launcher.launch("image/*") // Abre selector de imágenes, permitiendo cambiar la imagen al hacerle click encima de la foto
-            }) {
-            if (imageUri != null) {
-                //Mostrar imagen seleccionada
-                Image(
-                    painter = rememberAsyncImagePainter(imageUri), //Renderiza la imagen directamente desde su URI
-                    contentDescription = "Avatar seleccionado",
-                    modifier = Modifier.fillMaxSize()
+    Scaffold(
+        // Barra Superior con título y acciones
+        topBar = {
+            TopAppBar(
+                // Boton para volver a la pantalla anterior
+                title = {
+                    Row (verticalAlignment = Alignment.CenterVertically) {
+                        IconButton(onClick = { navController.popBackStack() }) {
+                            Icon(
+                                Icons.Filled.ArrowBackIosNew,
+                                contentDescription = "Volver Atras"
+                            )
+                        }
+                    }
+                },
+                // Colores para la barra superior
+                colors = TopAppBarDefaults.topAppBarColors(
+                    containerColor = Color(0xFF242424),  // Color de fondo
+                    titleContentColor = Color.White,      // Color del texto
+                    actionIconContentColor = Color.White  // Color de los iconos de acción
+                ),
+                actions = {
+                    // Botón para ir a Configuración
+                    IconButton(onClick = { navController.navigate(route = AppScreens.SettingsScreen.route) }) {
+                        Icon(
+                            Icons.Filled.Settings,
+                            contentDescription = "Seccion de Settings"
+                        )
+                    }
+                }
+            )
+        },
+
+        // Implementacion para Material3:
+        // Barra inferior con navegacion principal
+        bottomBar = {
+            // Navegacion inferior con iconos
+            NavigationBar(
+                containerColor = Color(0xFF242424),
+                contentColor = Color.White
+            ) {
+
+                // Esquema de color para los diferentes estados de los botones
+                val navBarItemColors = NavigationBarItemDefaults.colors(
+                    selectedIconColor = Color.White,      // Ícono seleccionado
+                    unselectedIconColor = Color.White,     // Ícono no seleccionado
+                    selectedTextColor = Color.White,      // Texto seleccionado
+                    unselectedTextColor = Color.White,      // Texto no seleccionado
+                    indicatorColor = Color.Transparent     // Quitar el recuadro
                 )
-            } else {
-                //Mostrar imagen por defecto
-                Icon(
-                    imageVector = Icons.Default.AccountCircle,
-                    contentDescription = "Avatar por defecto",
-                    modifier = Modifier.fillMaxSize(),
-                    tint = Color.Gray
+
+                // Boton de Home o inicio
+                NavigationBarItem(
+                    icon = { Icon(Icons.Filled.Home, contentDescription = "Inicio") },
+                    label = { Text("Inicio")},
+                    colors = navBarItemColors,
+                    selected = true,
+                    onClick = { navController.navigate(route = AppScreens.HomeScreen.route) }
+                )
+                // Boton de Favoritos
+                NavigationBarItem(
+                    icon = { Icon(Icons.Filled.Favorite, contentDescription = "Favoritos") },
+                    label = { Text("Favoritos")},
+                    colors = navBarItemColors,
+                    selected = true,
+                    onClick = { navController.navigate(AppScreens.FavoritesScreen.route)}
+                )
+                // Boton para vender
+                NavigationBarItem(
+                    icon = { Icon(Icons.Filled.Shop, contentDescription = "Vender") },
+                    label = { Text("Vender")},
+                    colors = navBarItemColors,
+                    selected = true,
+                    onClick = { /* TODO: Implementar navegacion */ }
+                )
+                // Boton de Categorias
+                NavigationBarItem(
+                    icon = { Icon(Icons.Filled.Menu, contentDescription = "Categorias") },
+                    label = { Text("Categorias")},
+                    colors = navBarItemColors,
+                    selected = true,
+                    onClick = { navController.navigate(AppScreens.CategoryScreen.route) }
                 )
             }
         }
-
-        Spacer(Modifier.height(12.dp))
-
-        //Info personal
-        //Campos editables: nombre completo, teléfono, domicilio y ciudad
-        EditableProfileItem("Nombre completo", fullName, isNameValid) { fullName = it }
-        EditableProfileItem("Correo electrónico", email, isEmailValid) { email = it }
-        EditableProfileItem("Teléfono", phoneNumber, isPhoneValid) { phoneNumber = it }
-        EditableProfileItem("Domicilio", address, isAddressValid) { address = it }
-        EditableProfileItem("Ciudad", city, isCityValid) { city = it }
-
-        Spacer(Modifier.height(24.dp))
-
-        //Texto clickeable para cambiar la contraseña
-        Text("Cambiar contraseña", color = Color.Blue, modifier = Modifier.clickable {
-            showPasswordDialog = true
-        })
-        //Texto clickeable para ver mis productos
-        Text("Mis productos", color = Color.Blue, modifier = Modifier.clickable {
-            Toast.makeText(context, "Mis productos (futuro)", Toast.LENGTH_SHORT).show()
-        })
-        //Texto clickeable para eliminar la cuenta
-        Text("Eliminar cuenta", color = Color.Red, modifier = Modifier.clickable {
-            Toast.makeText(context, "Eliminar cuenta (futuro)", Toast.LENGTH_SHORT).show()
-        })
-
-        Spacer(Modifier.height(32.dp))
-
-        Button(
-            onClick = {
-                Toast.makeText(context, "Cambios guardados correctamente", Toast.LENGTH_SHORT).show()
-            },
-            enabled = isFormValid, //Para poder clickear el boton debe estar previamente validado el formulario
+    ) { paddingValues ->
+        Column(
             modifier = Modifier
-                .fillMaxWidth()
-                .padding(vertical = 8.dp)
+                .fillMaxSize()
+                .padding(paddingValues)
         ) {
-            Text("Guardar cambios")
-        }
-
-        //Botón para cerrar sesión
-        Button(
-            onClick = { showLogoutDialog = true }, //Te redirige a la pantalla de login
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Text("Cerrar sesión")
-        }
-
-        // 🔐 Dialog cambiar contraseña
-        if (showPasswordDialog) {
-            ChangePasswordDialog(
-                newPassword = newPassword,
-                repeatPassword = repeatPassword,
-                onPasswordChange = { newPassword = it },
-                onRepeatChange = { repeatPassword = it },
-                //Esta accion ocurre cuando el usuario presiona "cancelar" o cierra el diálogo: se oculta el diálogo y se limpian los campos de contraseña
-                onDismiss = {
-                    showPasswordDialog = false
-                    newPassword = ""
-                    repeatPassword = ""
-                },
-                //Esta accion ocurre cuando el usuario presiona "confirmar: se limpian los campos y se cierra el diálogo
-                onConfirm = {
-                    Toast.makeText(context, "Contraseña cambiada", Toast.LENGTH_SHORT).show()
-                    newPassword = ""
-                    repeatPassword = ""
-                    showPasswordDialog = false
-                }
-            )
-        }
-
-        // 🚪 Dialog cerrar sesión
-        if (showLogoutDialog) {
-            LogoutConfirmationDialog(
-                //En el caso que se seleccione que si se desea cerrar sesión, navega hacia la pantalla de login, evitando que se pueda volver hacia esta pantalla de perfil
-                onConfirm = {
-                    Toast.makeText(context, "Sesión cerrada", Toast.LENGTH_SHORT).show()
-                    showLogoutDialog = false
-                    navController.navigate(AppScreens.LoginScreen.route) {  //Redirige a la pantalla de login
-                        //Quita todas las pantallas hasta la especificada (la de perfil)
-                        popUpTo(AppScreens.ProfileScreen.route) { inclusive = true }  //inclusive = true --> Remueve la pantalla actual ProfileScreen
-                        launchSingleTop = true //Evita que se creen múltiples instancias si ya está en el top del stack
+            Column(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(16.dp)
+                    .verticalScroll(rememberScrollState()),
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
+                // Imagen de perfil clickeable
+                Box(
+                    modifier = Modifier
+                        .size(100.dp)
+                        .clip(CircleShape) //Recorta el contenido de la box formando un circulo
+                        .clickable {
+                            //Solo mostrar archivos que comiencen con image/, como image/jpeg, image/png, etc. (de tipo MIME)
+                            launcher.launch("image/*") // Abre selector de imágenes, permitiendo cambiar la imagen al hacerle click encima de la foto
+                        }) {
+                    if (imageUri != null) {
+                        //Mostrar imagen seleccionada
+                        Image(
+                            painter = rememberAsyncImagePainter(imageUri), //Renderiza la imagen directamente desde su URI
+                            contentDescription = "Avatar seleccionado",
+                            modifier = Modifier.fillMaxSize()
+                        )
+                    } else {
+                        //Mostrar imagen por defecto
+                        Icon(
+                            imageVector = Icons.Default.AccountCircle,
+                            contentDescription = "Avatar por defecto",
+                            modifier = Modifier.fillMaxSize(),
+                            tint = Color.Gray
+                        )
                     }
-                },
-                //En el caso que se descarte la opción, es decir, se seleccione que no se desea cerrar sesión, quita el dialog de cerrar sesión de la pantalla
-                onDismiss = { showLogoutDialog = false }
-            )
+                }
+
+                Spacer(Modifier.height(12.dp))
+
+                //Info personal
+                //Campos editables: nombre completo, teléfono, domicilio y ciudad
+                EditableProfileItem("Nombre completo", fullName, isNameValid) { fullName = it }
+                EditableProfileItem("Correo electrónico", email, isEmailValid) { email = it }
+                EditableProfileItem("Teléfono", phoneNumber, isPhoneValid) { phoneNumber = it }
+                EditableProfileItem("Domicilio", address, isAddressValid) { address = it }
+                EditableProfileItem("Ciudad", city, isCityValid) { city = it }
+
+                Spacer(Modifier.height(24.dp))
+
+                //Texto clickeable para cambiar la contraseña
+                Text("Cambiar contraseña", color = Color.Blue, modifier = Modifier.clickable {
+                    showPasswordDialog = true
+                })
+                //Texto clickeable para ver mis productos
+                Text("Mis productos", color = Color.Blue, modifier = Modifier.clickable {
+                    Toast.makeText(context, "Mis productos (futuro)", Toast.LENGTH_SHORT).show()
+                })
+                //Texto clickeable para eliminar la cuenta
+                Text("Eliminar cuenta", color = Color.Red, modifier = Modifier.clickable {
+                    Toast.makeText(context, "Eliminar cuenta (futuro)", Toast.LENGTH_SHORT).show()
+                })
+
+                Spacer(Modifier.height(32.dp))
+
+                Button(
+                    onClick = {
+                        Toast.makeText(
+                            context,
+                            "Cambios guardados correctamente",
+                            Toast.LENGTH_SHORT
+                        ).show()
+                    },
+                    enabled = isFormValid, //Para poder clickear el boton debe estar previamente validado el formulario
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(vertical = 8.dp)
+                ) {
+                    Text("Guardar cambios")
+                }
+
+                //Botón para cerrar sesión
+                Button(
+                    onClick = { showLogoutDialog = true }, //Te redirige a la pantalla de login
+                    modifier = Modifier.fillMaxWidth()
+                ) {
+                    Text("Cerrar sesión")
+                }
+
+                // 🔐 Dialog cambiar contraseña
+                if (showPasswordDialog) {
+                    ChangePasswordDialog(
+                        newPassword = newPassword,
+                        repeatPassword = repeatPassword,
+                        onPasswordChange = { newPassword = it },
+                        onRepeatChange = { repeatPassword = it },
+                        //Esta accion ocurre cuando el usuario presiona "cancelar" o cierra el diálogo: se oculta el diálogo y se limpian los campos de contraseña
+                        onDismiss = {
+                            showPasswordDialog = false
+                            newPassword = ""
+                            repeatPassword = ""
+                        },
+                        //Esta accion ocurre cuando el usuario presiona "confirmar: se limpian los campos y se cierra el diálogo
+                        onConfirm = {
+                            Toast.makeText(context, "Contraseña cambiada", Toast.LENGTH_SHORT)
+                                .show()
+                            newPassword = ""
+                            repeatPassword = ""
+                            showPasswordDialog = false
+                        }
+                    )
+                }
+
+                // 🚪 Dialog cerrar sesión
+                if (showLogoutDialog) {
+                    LogoutConfirmationDialog(
+                        //En el caso que se seleccione que si se desea cerrar sesión, navega hacia la pantalla de login, evitando que se pueda volver hacia esta pantalla de perfil
+                        onConfirm = {
+                            Toast.makeText(context, "Sesión cerrada", Toast.LENGTH_SHORT).show()
+                            showLogoutDialog = false
+                            navController.navigate(AppScreens.LoginScreen.route) {  //Redirige a la pantalla de login
+                                //Quita todas las pantallas hasta la especificada (la de perfil)
+                                popUpTo(AppScreens.ProfileScreen.route) {
+                                    inclusive = true
+                                }  //inclusive = true --> Remueve la pantalla actual ProfileScreen
+                                launchSingleTop =
+                                    true //Evita que se creen múltiples instancias si ya está en el top del stack
+                            }
+                        },
+                        //En el caso que se descarte la opción, es decir, se seleccione que no se desea cerrar sesión, quita el dialog de cerrar sesión de la pantalla
+                        onDismiss = { showLogoutDialog = false }
+                    )
+                }
+            }
         }
     }
 }
